backend/redflags/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
slowapi
drain3